Fix internal link detection
Links that point to internal sites on a different subdomain from hosted app are currently considered 'internal' links because the root domain (as returned by wurl) is the same. 

This change makes sure the subdomain is considered. Downside is that foo.bar.com will be considered different from bar.com - which is what should happen technically, though not practically. Maybe this should be a flag?