Get the zoomFactor from the webContents before changing it
This fixes an issue that occurs on app relaunch.  The main window remembers the zoom level it was at across launches, but after relaunching when you try to adjust the zoom level, the adjustment happens relative to the original options.zoom value instead of the restored zoom value, causing an unexpected jump/reset of zoom level.