Always open external links externally
When Google Calendar performs the about:blank redirect technique, the redirect that it performs causes the onNewWIndow handler to be called with a 'dispostion' value of 'foreground-tab'. Because the disposition check happens before the linkIsInternal() check happens, it is opened in in a tab instead of in the external browser. This change moves the linkIsInternal() check earlier so that external links are always opened externally, regardless of disposition.