Always open external links externally
The tab feature introduced by #579 included a change that checks the 'disposition' parameter and conditionally creates tabs, and that check was placed prior to the check to see if the URL is internal. This change moves the linkIsInternal() check earlier so that external links are always opened externally, regardless of disposition.