Make `getCurrentUrl` actually return something
When the getCurrentUrl function was refactored to make use of the `withFocusedWindow` function in ac99c6424d7326ad7d1da86922866e6776ce3220, it stopped returning a value and broke the "Copy Current URL" feature (#633).

This change restores the original behavior of the getCurrentURL function and makes the "Copy Current URL" feature functional again.