Fix Docker build for issue #1501
The changes in the Dockerfile are related to updating the base image, installing dependencies and setting up the environment for building nativefier apps using Docker.

On the first change, the base image is updated from node:lts-alpine to node:16.3.0-alpine which means that the image is now based on Alpine Linux 3.13 with Node.js version 16.3.0.

In the second change, the npm packages prefix is set to $NPM_PACKAGES and the rimraf and typescript packages are installed globally.

In the third change, the @types/electron package is installed as a dev dependency, npm install command is run, and npm run build is executed.

In the change in the npm-shrinkwrap.json, the version of electron-packager is updated from 15.5.1 to 17.1.1 and @types/asar and @types/electron-packager were added.

In the change in the package.json, the version of electron-packager is updated from 15.5.1 to 17.1.1.

These changes are likely to make sure that the image is using the correct version of the packages and also to be compatible with the latest version of electron-packager and asar.